refactor: move `documentsToSelectChoices` to `helper/handlebars` module
