+++ module/applications/actor/character-generation-prompt.mjs
@@ -1,4 +1,4 @@
-import { documentsToSelectChoices, filterObject } from "../../utils.mjs";
+import { filterObject } from "../../utils.mjs";
 import CharacterGenerationData from "../../data/other/character-generation.mjs";
 import ItemEd from "../../documents/item.mjs";
 import ApplicationEd from "../api/application.mjs";
@@ -7,6 +7,7 @@ import * as LEGEND from "../../config/legend.mjs";
 import * as DOCUMENT_DATA from "../../config/document-data.mjs";
 import * as SYSTEM from "../../config/system.mjs";
 import { getAllDocuments } from "../../helper/document.mjs";
+import { documentsToSelectChoices } from "../../helper/handlebars.mjs";
 
 
 /**
