+++ module/helper/handlebars.mjs
@@ -1,11 +1,10 @@
-import getDice from "./dice/step-tables.mjs";
-import { linkForUuidSync } from "./utils.mjs";
-import * as LEGEND from "./config/legend.mjs";
+import getDice from "../dice/step-tables.mjs";
+import { linkForUuidSync } from "../utils.mjs";
+import * as LEGEND from "../config/legend.mjs";
 
 /**
  * @module handlebar-helpers - Provides custom Handlebars helpers for the application.
  */
-
 
 /**
  * Checks if a collection has any items.
@@ -174,4 +173,20 @@ export default function registerHandlebarHelpers() {
     "ed-wrapInArray":            wrapInArray,
   } );
 
+}
+
+/**
+ * Takes an array of documents and returns an object that can be used by Foundry's
+ * {@link selectOptions} Handlebar helper as choices. The keys are a document's
+ * UUID, the values it's name, which is rendered as representation in the HTML.
+ * @param {foundry.abstract.Document[]} documents An array of documents that should
+ * be the choices.
+ * @returns {{}} An object in the form of the `choices` parameter of the
+ * {@link selectOptions} Handlebar helper.
+ */
+export function documentsToSelectChoices( documents ) {
+  return documents.reduce(
+    ( obj, doc ) => ( { ...obj, [doc.uuid]: doc.name } ),
+    {}
+  );
 }
